# infrastructure/terraform/openapi.yaml
openapi: 3.0.3
info:
  title: Donation Platform API
  description: |
    Comprehensive API for the gamified donation platform with A2A agent integration.
    
    Features:
    - Donation processing with real-time gamification
    - Points and achievements system
    - A2A agent coordination for optimization
    - Location-based rewards and hotspots
    - Real-time analytics and leaderboards
    
  version: 1.0.0
  contact:
    name: Donation Platform Team
    email: support@donationplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: ${donation_service_url}
    description: Donation Service
  - url: ${points_service_url}
    description: Points & Achievements Service
  - url: ${donor_agent_url}
    description: Donor Engagement Agent (A2A)
  - url: ${charity_agent_url}
    description: Charity Optimization Agent (A2A)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
...
security:
  - BearerAuth: []
paths:
...
  /a2a/invoke:
    post:
      summary: Invoke agent skill
      description: A2A protocol endpoint for invoking agent skills
      tags:
        - A2A Protocol
      security:
        - BearerAuth: []
      parameters:
        - name: A2A-Protocol-Version
          in: header
          required: true
          schema:
            type: string
            example: "0.3.0"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - skill
                - params
              properties:
                id:
                  type: string
                  description: Unique message ID
                skill:
                  type: string
                  description: Name of skill to invoke
                params:
                  type: object
                  description: Parameters for the skill
                timestamp:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Skill executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  result:
                    type: object
                    description: Skill execution result
                  timestamp:
                    type: string
                    format: date-time
                  agent:
                    type: string
                    description: Name of the agent that processed the request
        '404':
          description: Skill not found
        '500':
          description: Skill execution failed

  # WebSocket Endpoints (documented for reference)
...